playground/requirements.txt
# CosyVoice TTS Requirements
dashscope>=1.14.0
python-dotenv>=1.0.0
